+++ buildout/src/Euphorie/docs/changes.rst
@@ -1,8 +1,14 @@
 Changelog
 =========
 
-2.1 - Unreleased
+2.2 - Unreleased
 ----------------
+
+- ...
+
+
+2.1 - November 6, 2010
+----------------------
 
 Feature changes
 ~~~~~~~~~~~~~~~